requirements.txt
anthropic>=0.40.0
websockets>=12.0
requests>=2.31.0
python-dotenv>=1.0.0
py-clob-client>=0.18.0
